refactor: move batch methods into dataloader service
