+++ src/votes/votes.service.ts
@@ -1,22 +1,16 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { FindOptionsWhere, Repository } from 'typeorm';
-import { Proposal } from '../proposals/models/proposal.model';
 import { ProposalsService } from '../proposals/proposals.service';
 import { CreateVoteInput } from './models/create-vote.input';
 import { UpdateVoteInput } from './models/update-vote.input';
 import { Vote } from './models/vote.model';
-
-type ProposalWithVoteCount = Proposal & { voteCount: number };
 
 @Injectable()
 export class VotesService {
   constructor(
     @InjectRepository(Vote)
     private repository: Repository<Vote>,
-
-    @InjectRepository(Proposal)
-    private proposalsRepository: Repository<Proposal>,
 
     private proposalsService: ProposalsService,
   ) {}
@@ -27,26 +21,6 @@ export class VotesService {
 
   async getVotes(where?: FindOptionsWhere<Vote>) {
     return this.repository.find({ where });
-  }
-
-  async getVoteCountBatch(proposalIds: number[]) {
-    const proposals = (await this.proposalsRepository
-      .createQueryBuilder('proposal')
-      .leftJoinAndSelect('proposal.votes', 'vote')
-      .loadRelationCountAndMap('proposal.voteCount', 'proposal.votes')
-      .select(['proposal.id'])
-      .whereInIds(proposalIds)
-      .getMany()) as ProposalWithVoteCount[];
-
-    return proposalIds.map((id) => {
-      const proposal = proposals.find(
-        (proposal: Proposal) => proposal.id === id,
-      );
-      if (!proposal) {
-        return new Error(`Could not load vote count for proposal: ${id}`);
-      }
-      return proposal.voteCount;
-    });
   }
 
   async createVote(voteData: CreateVoteInput, userId: number) {
